GameScore tests fixed. Function created.
